requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
lxml==4.8.0
MarkupSafe==2.1.1
multitasking==0.0.10
mysql-connector==2.2.9
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.1
yfinance==0.1.70
zipp==3.8.0
